change f32 to f64

+++ src/linear_algebra.rs
@@ -4,24 +4,24 @@ use wasm_bindgen::prelude::*;
 #[wasm_bindgen]
 #[derive(Debug, Clone, Copy)]
 pub struct V2 {
-    pub x : f32, 
-    pub y : f32,
+    pub x : f64, 
+    pub y : f64,
 }
 
 #[wasm_bindgen]
 impl V2 {
     #[wasm_bindgen(constructor)]
-    pub fn new(x: f32, y: f32) -> Self {
+    pub fn new(x: f64, y: f64) -> Self {
         Self { x, y }
     }
 
-    pub fn rotate(&self, angle : f32) -> Self {
+    pub fn rotate(&self, angle : f64) -> Self {
         let xnew = self.x * angle.cos() - self.y * angle.sin();
         let ynew = self.x * angle.sin() + self.y * angle.cos();
         Self { x:xnew, y:ynew }
     }
 
-    pub fn modulus(&self) -> f32 {
+    pub fn modulus(&self) -> f64 {
         if self.x == 0.0 { return self.y; }
         if self.y == 0.0 { return self.x; }
         (self.x * self.x + self.y * self.y).sqrt()
@@ -31,7 +31,7 @@ impl V2 {
         self * (1.0/self.modulus())
     }
 
-    pub fn angle(&self) -> f32 {
+    pub fn angle(&self) -> f64 {
         self.y.atan2(self.x)
     }
 }
@@ -48,9 +48,9 @@ impl ops::Sub for V2 {
         V2 { x:self.x - other.x, y:self.y - other.y}
     }
 }
-impl ops::Mul<f32> for V2 {
+impl ops::Mul<f64> for V2 {
     type Output = Self;
-    fn mul(self, scalar: f32) -> Self::Output {
+    fn mul(self, scalar: f64) -> Self::Output {
         V2 { x:self.x * scalar, y:self.y * scalar}
     }
 }
